move datalogger and command service into board b/c they compile for target
