+++ board/app/Cargo.toml
@@ -17,4 +17,4 @@ rtt-target = "0.4.0"
 embedded-alloc = "0.5.0"
 
 rriv-0-4 = { path = "../rriv-0-4" }
-rriv-datalogger = { path = "../../src/rriv-datalogger" }
+datalogger = { path = "../datalogger" }
