+++ board/command_service/Cargo.toml
@@ -1,14 +1,12 @@
 [package]
-name = "rriv-datalogger"
+name = "command_service"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rriv-0-4 = { path = "../../board/rriv-0-4" }
-control_interface = { path = "../control_interface" }
+rriv-0-4 = { path = "../rriv-0-4" }
+control_interface = { path = "../../src/control_interface" }
 hashbrown = "0.14"
 cortex-m = { version = "0.7.6", features = ["critical-section-single-core"] }
-
-
