+++ board/datalogger/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "datalogger"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+cortex-m = { version = "0.7.6", features = ["critical-section-single-core"] }
+rriv-0-4 = { path = "../rriv-0-4" }
+command_service = { path = "../command_service" }
