+++ board/lib/build.rs
@@ -16,6 +16,8 @@ use std::fs::File;
 use std::io::Write;
 use std::path::PathBuf;
 
+
+#[allow(dead_code)]
 fn macos_copy_memory_x() {
     // Put `memory.x` in our output directory and ensure it's
     // on the linker search path.
