+++ board/lib/src/lib.rs
@@ -4,10 +4,30 @@
 #![no_main]
 
 extern crate panic_halt;
+use core::{prelude::rust_2024::*, u8};
+use cortex_m_rt::entry;
+use rtt_target::rtt_init_print;
 
 pub mod prelude;
 
-// TODO:: import the libraries needed and update these FFI functions
+extern crate rriv_0_4;
+use rriv_0_4::Board;
+
+extern crate datalogger;
+use datalogger::DataLogger;
+
+#[entry]
+fn main() -> ! {
+    prelude::init();
+    rtt_init_print!();
+
+    let board = Board::new();
+    let mut datalogger = DataLogger::new(board);
+    loop {
+        datalogger.run_loop_iteration();
+    }
+}
+// // TODO:: import the libraries needed and update these FFI functions
 // #[no_mangle]
 // pub unsafe extern "C" fn rust_serial_interface_new() -> *mut c_void {
 //     let board = Board::init();
