+++ board/rriv-0-4/src/lib.rs
@@ -113,7 +113,6 @@ impl Board {
     }
 }
 
-
 #[interrupt]
 unsafe fn USART2() {
     cortex_m::interrupt::free(|cs| {
